Grouping and moving UI and model classes to their own files and groups
