+++ EssentialFeediOS/Feed Image Loader/FeedImageDataLoader.swift
@@ -1,0 +1,18 @@
+//
+//  FeedImageDataLoader.swift
+//  EssentialFeediOS
+//
+//  Created by Macbook Pro on 18/9/23.
+//
+
+import Foundation
+
+public protocol FeedImageDataLoaderTask {
+    func cancel()
+}
+
+public protocol FeedImageDataLoader {
+    typealias Result = Swift.Result<Data, Error>
+    
+    func loadImageData(from url: URL, completion: @escaping (Result) -> Void) -> FeedImageDataLoaderTask
+}
